requirements.txt
gunicorn==20.1.0
uvicorn==0.17.6
fastapi==0.78.0

aiohttp==3.8.1
pytest-asyncio==0.18.3
coverage==6.4.1